Mudança de variáveis

+++ Models/DadosContrato/Fatura.cs
@@ -19,7 +19,7 @@ public class Fatura
     public DateTime DataVencimento { get; set; }
 
     [Required(ErrorMessage = "O valor total da fatura é obrigatório")]
-    public int ValorTotal { get; set; }
+    public decimal ValorTotal { get; set; }
 
     [Required(ErrorMessage = "O status da fatura é obrigatório")]
     [MaxLength(8, ErrorMessage = "Limite de caracteres excedido")]
